move out replanningModule
