+++ contribs/freight/src/main/java/org/matsim/freight/logistics/examples/multipleChains/ExampleTwoLspsGroceryDeliveryMultipleChains.java
@@ -58,6 +58,7 @@ import org.matsim.core.replanning.GenericPlanStrategyImpl;
 import org.matsim.core.replanning.selectors.BestPlanSelector;
 import org.matsim.core.replanning.selectors.ExpBetaPlanSelector;
 import org.matsim.core.replanning.selectors.GenericWorstPlanForRemovalSelector;
+import org.matsim.core.replanning.selectors.KeepSelected;
 import org.matsim.core.scenario.ScenarioUtils;
 import org.matsim.freight.carriers.*;
 import org.matsim.freight.carriers.analysis.CarriersAnalysis;
@@ -202,7 +203,10 @@ final class ExampleTwoLspsGroceryDeliveryMultipleChains {
                                 () -> {
                                   LSPStrategyManager strategyManager = new LSPStrategyManagerImpl();
                                   strategyManager.addStrategy(new GenericPlanStrategyImpl<>(new ExpBetaPlanSelector<>(new ScoringConfigGroup())), null, 1);
-                                  strategyManager.addStrategy(RandomShiftingStrategyFactory.createStrategy(), null, 4);
+
+									GenericPlanStrategyImpl<LSPPlan, LSP> strategy = new GenericPlanStrategyImpl<>(new KeepSelected<>());
+									strategy.addStrategyModule(new LspRandomShipmentShiftingModule());
+									strategyManager.addStrategy(strategy, null, 4);
                                   strategyManager.setMaxPlansPerAgent(5);
                                   strategyManager.setPlanSelectorForRemoval(new GenericWorstPlanForRemovalSelector<>());
                                   return strategyManager;
